requirements.txt
langchain 
langchain-community
python-docx 
pypdf2 
pdfplumber 
pdf2image 
pytesseract 
chromadb 
sentence-transformers 
unstructured
